integration_test: split tests into separate files
