+++ integration_test/info_test.go
@@ -1,0 +1,39 @@
+// Copyright 2019 The Kubernetes Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package integrationtest
+
+import "testing"
+
+func TestKrewInfo(t *testing.T) {
+	skipShort(t)
+
+	test, cleanup := NewTest(t)
+	defer cleanup()
+
+	test.WithIndex().Krew("info", validPlugin).RunOrFail()
+}
+
+func TestKrewInfoInvalidPlugin(t *testing.T) {
+	skipShort(t)
+
+	test, cleanup := NewTest(t)
+	defer cleanup()
+
+	plugin := "invalid-plugin"
+	err := test.WithIndex().Krew("info", plugin).Run()
+	if err == nil {
+		t.Errorf("Expected `krew info %s` to fail", plugin)
+	}
+}
