+++ integration_test/list_test.go
@@ -1,0 +1,39 @@
+// Copyright 2019 The Kubernetes Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package integrationtest
+
+import (
+	"bytes"
+	"testing"
+)
+
+func TestKrewList(t *testing.T) {
+	skipShort(t)
+
+	test, cleanup := NewTest(t)
+	defer cleanup()
+
+	initialList := test.WithIndex().Krew("list").RunOrFailOutput()
+	if bytes.Contains(initialList, []byte(validPlugin)) {
+		t.Errorf("%q should initially not be installed", validPlugin)
+	}
+
+	test.Krew("install", validPlugin).RunOrFail()
+
+	eventualList := test.Krew("list").RunOrFailOutput()
+	if !bytes.Contains(eventualList, []byte(validPlugin)) {
+		t.Errorf("%q should eventually be installed", validPlugin)
+	}
+}
